Make Tier and DataCenter properties of NFigStore
Providing the functionality to get setting for a tier other than
the one your app is currently running on is misleading. The
settings wouldn't be accurate if that tier has overrides. Removing
tier and dataCenter as parameters for getting or subscribing to
app settings makes the interface simpler and reduces the chances
of misleading users.