Replace BySettingDictionary
I really needed two different dictionaries. One to implement
IReadOnlyDictionary<string, T>, and the other to implement
IReadOnlyDictionary<string, IReadOnlyList<T>>, but I still wanted both
to be easily instantiated via a collection of T.

The solution was to create an implementation which supported either
representation. See BySettingBase.cs for details. The functionality is
exposed by the BySetting<TValue> and ListBySetting<TValue>.