Add DefaultValueBaseAttribute to replace old attribute
Previously, the DefaultSettingValueAttribute was the base class for
defining default value attributes. However, things need to be a little
bit more flexible in order to support sub-apps well. So the new
attribute simply has a method which allows the user to return as many
default values as they want from a single attribute.

This will require slightly more implementation on the consumer's side,
but is a lot more flexible.

Related to #19