Remove UnthrownStackTrace property from NFigException
It was only used for the InvalidOverridesException, and no one really
cares about the stack trace on that. The StackTrace(bool) constructor we
need in order to generate a stack trace without throwing is not available
in .NET Standard prior to 2.0, so it's easiest to just get rid of it.