Refactor how default values are generated
The DefaultValue class was not a good fit for generating defaults from the
DefaultValueBaseAttribute because the Value property was a string. In many
cases, users will provide non-string literals (such as integers).

NFig normalizes everything to a string via the setting's converter, but
passing the converter into the attribute is clunky and impractical. It's
easier for the attribute to simply generate another type where the value
property is of type object, and let NFig convert it into a proper
DefaultValue object.