Refactor how sub-apps are registered
There is now only one method for registering sub-apps, rather than three.
It allows you to register more than one at a time.

The method for setting/updating metadata for the sub-app in the store was
also refactored. It's really better to allow multiple sub-apps to be
updated at the same time so that we only have to trigger one notification
that metadata has changed.