Remove most async methods from Store and clients
The plan is to have the Store cache data, and most methods will simply
read from the cache instead of performing any I/O. The cache will
automatically update in most cases, but we'll also provide a way to force
it to update (not implemented yet).

I'm also starting over on the MemoryStore. In the implementation was a bit
muddy. I want to create a MockRedis class so that the MemoryStore can be a
more realistic test.